Move provenance data into own module
