+++ libs/@local/hash-graph-types/rust/src/knowledge/entity/mod.rs
@@ -1,3 +1,5 @@
+mod provenance;
+
 use std::{fmt, str::FromStr};
 
 use error_stack::Report;
@@ -10,6 +12,7 @@ use type_system::url::{BaseUrl, VersionedUrl};
 use utoipa::{openapi, ToSchema};
 use uuid::Uuid;
 
+pub use self::provenance::{EntityEditionProvenanceMetadata, EntityProvenanceMetadata};
 use crate::{
     account::{CreatedById, EditionCreatedById},
     knowledge::{
@@ -77,29 +80,6 @@ impl fmt::Display for DraftId {
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&self.0, fmt)
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-pub struct EntityEditionProvenanceMetadata {
-    pub created_by_id: EditionCreatedById,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-pub struct EntityProvenanceMetadata {
-    pub created_by_id: CreatedById,
-    pub created_at_transaction_time: Timestamp<TransactionTime>,
-    pub created_at_decision_time: Timestamp<DecisionTime>,
-    #[cfg_attr(feature = "utoipa", schema(nullable = false))]
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub first_non_draft_created_at_transaction_time: Option<Timestamp<TransactionTime>>,
-    #[cfg_attr(feature = "utoipa", schema(nullable = false))]
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub first_non_draft_created_at_decision_time: Option<Timestamp<DecisionTime>>,
-    pub edition: EntityEditionProvenanceMetadata,
 }
 
 /// The metadata of an [`Entity`] record.
